Fixed TopNSorted list. Small refactoring.
